pyproject.toml
[project]
name = "auth-microservice"
version = "1.0.0"
description = "Authentication microservice with hexagonal architecture"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "aiosqlite==0.19.0",
    "asyncpg==0.29.0",
    "psycopg2-binary==2.9.9",
    "pydantic[email]==2.5.0",
    "pydantic-settings==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.0",
    "pika==1.3.2",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-mock>=3.12.0",
    "httpx==0.25.1",
    "pytest-cov>=7.0.0",
    "aio-pika>=9.6.1",
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v -W ignore::RuntimeWarning"
filterwarnings = [
    "ignore::RuntimeWarning",
    "ignore::UserWarning",
    "ignore:Task was destroyed but it is pending::RuntimeWarning",
    "ignore:no running event loop::RuntimeWarning",
    "ignore:coroutine.*was never awaited::RuntimeWarning",
    "ignore:Exception ignored in::RuntimeWarning",
]

[tool.pylint.messages_control]
disable = [
    "missing-class-docstring",
    "missing-function-docstring",
    "missing-module-docstring",
]

[tool.pylint.format]
max-line-length = 88
